+++ mateusz_tunski/dollacop/imports/startup/server/index.js
@@ -1,6 +1,6 @@
 import { Meteor } from "meteor/meteor"
 
 import "/imports/api/Debts"
-import "./fixtures"
+import "./seeds"
 
 Meteor.startup(() => {})
